backend/pyproject.toml
[project]
name = "catwiki-backend"
version = "1.0.0"
description = "CatWiki 后端 API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "gunicorn>=21.2.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "email-validator>=2.1.0",
    "sqlalchemy>=2.0.23",
    "asyncpg>=0.29.0",
    "greenlet>=3.0.0",
    "alembic>=1.13.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.0.0",
    "minio>=7.2.0",
    "langchain-postgres>=0.0.16",
    "langchain-openai>=1.1.7",
    "langchain-core>=1.2.6",
    "psycopg2-binary>=2.9.9",
    "langchain-text-splitters>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "httpx>=0.25.2",
    "black>=23.11.0",
    "ruff>=0.1.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = [
    "E501",   # 行长度由 black 处理
    "N802",   # 允许大写属性名（如 DATABASE_URL 环境变量风格）
    "N805",   # 允许 cls 作为 declared_attr 的第一个参数
    "N815",   # 允许 Pydantic 字段使用 camelCase（JSON 兼容）
    "UP046",  # 保持 Pydantic Generic 当前写法
    "N818",   # 允许 Exception 后缀（不强制使用 Error 后缀）
]

